refactor(modularization): Phase 3 Step 5b2b — collection_service move to KB domain + DI setters + legacy shim (#16)
Step 5b2b in the Phase 3 #1629 atomic sequence (msg=12672498 8-step
plan; sub-sub-order per msg=99183dff; Protocol+DI pattern per
msg=6ab7d211 Q1/Q2 + msg=896584ee QuotaInitOps parity).

**ports.py updates (msg=6ab7d211 Q2 + msg=896584ee):**
- Rename `MarketplaceCollectionOps._check_marketplace_access` →
  `check_marketplace_access` (drop `_` prefix; Phase 4
  marketplace_collection_service move structurally satisfies the
  public name at its canonical location).
- Add consumer-owned `QuotaOps` Protocol (`check_and_consume_quota` +
  `release_quota`). Phase 4 keeps quota_service legacy; Phase 5
  design-lock decides permanent home.

**collection_service move (aperag/service → KB domain):**
- `git mv` (rename preserved across tree walk even if git rewrites
  as new+modify due to large diff) body kept verbatim.
- G1 boundary pass: imports of `aperag.db.models` /
  `aperag.schema.view_models` / `aperag.service.*` rewritten to
  `aperag.domains.knowledge_base.db.models` /
  `aperag.domains.knowledge_base.schemas` /
  `aperag.domains.retrieval.schemas` / `aperag.schema.common` /
  sibling `aperag.domains.knowledge_base.service.collection_summary_service`.
- Cross-service dependencies (`marketplace_service`,
  `marketplace_collection_service`, `search_pipeline_service`,
  `quota_service`) reached via module-level DI singletons + Protocol
  accessor functions that raise `RuntimeError` when unwired, so
  missing wire-up fails loudly rather than limping along with `None`.
- `set_marketplace_ops` / `set_marketplace_collection_ops` /
  `set_search_pipeline_ops` / `set_quota_ops` public setters.

**legacy shim (aperag/service/collection_service.py):**
- Re-exports `CollectionService` + `collection_service` from domain
  path so pre-Phase-3 callers and FastAPI handlers still resolve.
- `_bootstrap_knowledge_base_collection_service_di()` wires the four
  DI setters at import time from the legacy service singletons; a
  small `_MarketplaceCollectionOpsAdapter` bridges the public
  `check_marketplace_access` Protocol name onto the legacy
  `_check_marketplace_access` method. Step 5b2c moves the wire-up
  into `aperag/app.py` startup; the shim+adapter collapse when
  Phase 4 marketplace_collection_service drops the `_` prefix.

**Boundary gates:**
- G1 green: `tests/unit_test/test_modularization_boundaries.py` all
  13 tests pass; AST walk scans the moved service and finds no
  legacy-aggregate imports (including the inline transaction closures
  that previously imported `aperag.db.models` and
  `aperag.service.quota_service`).
- Full unit suite green: 669 passed / 29 skipped.

Phase 3 total commit: 5b2b brings the `aperag.domains.knowledge_base`
surface to ports + schemas + db/models + collection_summary_service +
collection_service. 5b2c (app.py wire-up) + 5b3 (document_service
parity) + 5a (routes) remain.
