+++ aperag/service/collection_service.py
@@ -12,529 +12,68 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import logging
-from typing import List, Optional, Tuple
+"""Legacy re-export + transitional DI wire-up shim.
 
-import httpx
-from sqlalchemy.ext.asyncio import AsyncSession
+Phase 3 Step 5b2b moved the ``CollectionService`` body to
+``aperag.domains.knowledge_base.service.collection_service``. Pre-Phase-3
+callers and FastAPI handlers that still do
+``from aperag.service.collection_service import collection_service``
+continue to resolve through this module.
 
-from aperag.db import models as db_models
-from aperag.db.ops import AsyncDatabaseOps, async_db_ops
-from aperag.domains.knowledge_base.schemas import Collection
-from aperag.exceptions import ValidationException
-from aperag.schema import view_models
-from aperag.schema.utils import dumpCollectionConfig, parseCollectionConfig
-from aperag.schema.view_models import (
-    SearchResult,
-    SearchResultItem,
-    SearchResultList,
+The KB domain uses consumer-owned Protocols (Phase 3 Step 5b2a) for its
+four cross-domain dependencies: ``MarketplaceOps``,
+``MarketplaceCollectionOps``, ``SearchPipelineOps`` and ``QuotaOps``.
+Step 5b2c moves the concrete-instance wire-up into ``aperag/app.py``
+startup; until then this module bootstraps the DI setters at import
+time so the legacy import path keeps working unchanged.
+
+The ``MarketplaceCollectionOps`` Protocol uses the public method name
+``check_marketplace_access`` (msg=6ab7d211 Q2); the current concrete
+service still exposes it as ``_check_marketplace_access``. A thin
+adapter class bridges the two names until Phase 4 marketplace
+collection service drops the ``_`` prefix at its canonical location.
+Phase 6 cleanup removes this whole module once every caller is migrated
+to the domain path.
+"""
+
+from __future__ import annotations
+
+from aperag.domains.knowledge_base.service.collection_service import *  # noqa: F401, F403
+from aperag.domains.knowledge_base.service.collection_service import (  # noqa: F401
+    CollectionService,
+    collection_service,
+    set_marketplace_collection_ops,
+    set_marketplace_ops,
+    set_quota_ops,
+    set_search_pipeline_ops,
 )
-from aperag.service.collection_summary_service import collection_summary_service
-from aperag.service.marketplace_collection_service import marketplace_collection_service
-from aperag.service.marketplace_service import marketplace_service
-from aperag.service.search_pipeline_service import search_pipeline_service
-from aperag.utils.constant import QuotaType
-from aperag.views.utils import validate_source_connect_config
-from config.celery_tasks import collection_delete_task, collection_init_task
 
-logger = logging.getLogger(__name__)
 
+def _bootstrap_knowledge_base_collection_service_di() -> None:
+    """Import legacy service singletons and wire the Phase 3 Step 5b2a
+    consumer-owned Protocol setters. Kept internal and idempotent so a
+    re-import (e.g. test reset) does not double-wire. Step 5b2c replaces
+    this with explicit ``aperag/app.py`` startup wire-up."""
 
-class CollectionService:
-    """Collection service that handles business logic for collections"""
+    from aperag.service.marketplace_collection_service import marketplace_collection_service
+    from aperag.service.marketplace_service import marketplace_service
+    from aperag.service.quota_service import quota_service
+    from aperag.service.search_pipeline_service import search_pipeline_service
 
-    def __init__(self, session: AsyncSession = None):
-        # Use global db_ops instance by default, or create custom one with provided session
-        if session is None:
-            self.db_ops = async_db_ops  # Use global instance
-        else:
-            self.db_ops = AsyncDatabaseOps(session)  # Create custom instance for transaction control
+    class _MarketplaceCollectionOpsAdapter:
+        """Bridge the public ``check_marketplace_access`` Protocol name
+        onto the legacy ``_check_marketplace_access`` method. Phase 4
+        drops the ``_`` prefix at the concrete service; at that point
+        this adapter collapses to a straight passthrough (or is removed
+        once the service structurally satisfies the Protocol directly)."""
 
-    async def build_collection_response(self, instance: db_models.Collection) -> view_models.Collection:
-        """Build Collection response object for API return."""
-        return Collection(
-            id=instance.id,
-            title=instance.title,
-            description=instance.description,
-            type=instance.type,
-            status=getattr(instance, "status", None),
-            config=parseCollectionConfig(instance.config),
-            created=instance.gmt_created.isoformat(),
-            updated=instance.gmt_updated.isoformat(),
-        )
+        async def check_marketplace_access(self, user_id: str, collection_id: str) -> dict:
+            return await marketplace_collection_service._check_marketplace_access(user_id, collection_id)
 
-    async def create_collection(self, user: str, collection: view_models.CollectionCreate) -> view_models.Collection:
-        collection_config = collection.config
-        if collection.type != db_models.CollectionType.DOCUMENT:
-            raise ValidationException("collection type is not supported")
+    set_marketplace_ops(marketplace_service)
+    set_marketplace_collection_ops(_MarketplaceCollectionOpsAdapter())
+    set_search_pipeline_ops(search_pipeline_service)
+    set_quota_ops(quota_service)
 
-        is_validate, error_msg = validate_source_connect_config(collection_config)
-        if not is_validate:
-            raise ValidationException(error_msg)
 
-        # Create collection and consume quota in a single transaction
-        async def _create_collection_with_quota(session):
-            from aperag.service.quota_service import quota_service
-
-            # Check and consume quota within the transaction
-            await quota_service.check_and_consume_quota(user, "max_collection_count", 1, session)
-
-            # Create collection within the same transaction
-            config_str = dumpCollectionConfig(collection_config) if collection.config is not None else None
-
-            from aperag.db.models import Collection, CollectionStatus
-            from aperag.utils.utils import utc_now
-
-            instance = Collection(
-                user=user,
-                title=collection.title,
-                description=collection.description,
-                type=collection.type,
-                status=CollectionStatus.ACTIVE,
-                config=config_str,
-                gmt_created=utc_now(),
-                gmt_updated=utc_now(),
-            )
-            session.add(instance)
-            await session.flush()
-            await session.refresh(instance)
-
-            return instance
-
-        instance = await self.db_ops.execute_with_transaction(_create_collection_with_quota)
-
-        if collection.config.enable_summary:
-            await collection_summary_service.trigger_collection_summary_generation(instance)
-
-        # Initialize collection based on type
-        document_user_quota = await self.db_ops.query_user_quota(user, QuotaType.MAX_DOCUMENT_COUNT)
-        collection_init_task.delay(instance.id, document_user_quota)
-
-        return await self.build_collection_response(instance)
-
-    async def list_collections_view(
-        self, user_id: str, include_subscribed: bool = True, page: int = 1, page_size: int = 20
-    ) -> view_models.CollectionViewList:
-        """
-        Get user's collection list (lightweight view)
-
-        Args:
-            user_id: User ID
-            include_subscribed: Whether to include subscribed collections, default True
-            page: Page number
-            page_size: Page size
-        """
-        items = []
-
-        # 1. Get user's owned collections with marketplace info
-        owned_collections_data = await self.db_ops.query_collections_with_marketplace_info(user_id)
-
-        for row in owned_collections_data:
-            is_published = row.marketplace_status == "PUBLISHED"
-            items.append(
-                view_models.CollectionView(
-                    id=row.id,
-                    title=row.title,
-                    description=row.description,
-                    type=row.type,
-                    status=row.status,
-                    created=row.gmt_created,
-                    updated=row.gmt_updated,
-                    is_published=is_published,
-                    published_at=row.published_at if is_published else None,
-                    owner_user_id=row.user,
-                    owner_username=row.owner_username,
-                    subscription_id=None,  # Own collection, subscription_id is None
-                    subscribed_at=None,
-                )
-            )
-
-        # 2. Get subscribed collections if needed (optimized - no N+1 queries)
-        if include_subscribed:
-            try:
-                # Get subscribed collections data with all needed fields in one query
-                subscribed_collections_data, _ = await self.db_ops.list_user_subscribed_collections(
-                    user_id,
-                    page=1,
-                    page_size=1000,  # Get all subscriptions for now
-                )
-
-                for data in subscribed_collections_data:
-                    is_published = data["marketplace_status"] == "PUBLISHED"
-                    items.append(
-                        view_models.CollectionView(
-                            id=data["id"],
-                            title=data["title"],
-                            description=data["description"],
-                            type=data["type"],
-                            status=data["status"],
-                            created=data["gmt_created"],
-                            updated=data["gmt_updated"],
-                            is_published=is_published,
-                            published_at=data["published_at"] if is_published else None,
-                            owner_user_id=data["owner_user_id"],
-                            owner_username=data["owner_username"],
-                            subscription_id=data["subscription_id"],
-                            subscribed_at=data["gmt_subscribed"],
-                        )
-                    )
-            except Exception as e:
-                # If getting subscriptions fails, log and continue with owned collections
-                import logging
-
-                logger = logging.getLogger(__name__)
-                logger.warning(f"Failed to get subscribed collections for user {user_id}: {e}")
-
-        # 3. Sort by update time
-        items.sort(key=lambda x: x.updated or x.created, reverse=True)
-
-        # 4. Apply pagination
-        start_idx = (page - 1) * page_size
-        end_idx = start_idx + page_size
-        paginated_items = items[start_idx:end_idx]
-
-        return view_models.CollectionViewList(
-            items=paginated_items, pageResult=view_models.PageResult(total=len(items), page=page, page_size=page_size)
-        )
-
-    async def get_collection(self, user: str, collection_id: str) -> view_models.Collection:
-        from aperag.exceptions import CollectionNotFoundException
-
-        if not user:
-            await marketplace_service.validate_marketplace_collection(collection_id)
-            collection = await self.db_ops.query_collection_by_id(collection_id)
-        else:
-            collection = await self.db_ops.query_collection(user, collection_id)
-
-        if collection is None:
-            raise CollectionNotFoundException(collection_id)
-        return await self.build_collection_response(collection)
-
-    @staticmethod
-    def _embedding_identity(cfg) -> Optional[tuple]:
-        """Return a tuple that uniquely identifies an embedding model binding.
-
-        Returns None when no embedding is configured yet (e.g. freshly created
-        collection whose user has not filled it in), so first-time assignment is
-        still allowed.
-        """
-        if cfg is None:
-            return None
-        emb = getattr(cfg, "embedding", None)
-        if emb is None:
-            return None
-        model = getattr(emb, "model", None)
-        msp = getattr(emb, "model_service_provider", None)
-        clp = getattr(emb, "custom_llm_provider", None)
-        if not model and not msp and not clp:
-            return None
-        return (model, msp, clp)
-
-    def _reject_embedding_change(self, instance: db_models.Collection, update: view_models.CollectionUpdate) -> None:
-        """Raise ValidationException if the update tries to change the embedding binding."""
-        if update.config is None:
-            return
-        try:
-            old_cfg = parseCollectionConfig(instance.config) if instance.config else None
-        except Exception:
-            old_cfg = None
-
-        old_id = self._embedding_identity(old_cfg)
-        new_id = self._embedding_identity(update.config)
-
-        if old_id is None:
-            # First-time binding is allowed.
-            return
-        if new_id is None:
-            raise ValidationException(
-                "Embedding model of an existing collection cannot be cleared. "
-                "Keep the original embedding configuration or create a new collection."
-            )
-        if old_id != new_id:
-            old_model, _old_msp, _old_clp = old_id
-            new_model, _new_msp, _new_clp = new_id
-            raise ValidationException(
-                "Embedding model of an existing collection cannot be changed "
-                f"(current: {old_model!r}, requested: {new_model!r}). "
-                "Different embedding models produce vectors of different dimensions "
-                "and/or incompatible semantics, which would leave existing vectors "
-                "orphaned and break retrieval. Please create a new collection and "
-                "re-ingest your documents if a different model is required."
-            )
-
-    async def update_collection(
-        self, user: str, collection_id: str, collection: view_models.CollectionUpdate
-    ) -> view_models.Collection:
-        from aperag.exceptions import CollectionNotFoundException
-
-        # First check if collection exists
-        instance = await self.db_ops.query_collection(user, collection_id)
-        if instance is None:
-            raise CollectionNotFoundException(collection_id)
-
-        # Guardrail: embedding model/provider of an existing collection MUST NOT change.
-        # Different embedding models produce vectors with different dimensions and/or
-        # incompatible semantics. Silently switching will (a) leak orphan points in the
-        # previous global collection (under multi-tenant mode), (b) break retrieval
-        # recall, (c) in legacy mode even cause Qdrant to reject writes with a
-        # dimension mismatch error. Rejecting here is the least surprising behavior;
-        # users who truly need a different model should create a new collection and
-        # re-ingest their data.
-        self._reject_embedding_change(instance, collection)
-
-        # Direct call to repository method, which handles its own transaction
-        config_str = dumpCollectionConfig(collection.config)
-
-        updated_instance = await self.db_ops.update_collection_by_id(
-            user=user,
-            collection_id=collection_id,
-            title=collection.title,
-            description=collection.description,
-            config=config_str,
-        )
-
-        await collection_summary_service.trigger_collection_summary_generation(updated_instance)
-
-        if not updated_instance:
-            raise CollectionNotFoundException(collection_id)
-
-        return await self.build_collection_response(updated_instance)
-
-    async def delete_collection(self, user: str, collection_id: str) -> Optional[view_models.Collection]:
-        """Delete collection by ID (idempotent operation)
-
-        Returns the deleted collection or None if already deleted/not found
-        """
-        # Check if collection exists - if not, silently succeed (idempotent)
-        collection = await self.db_ops.query_collection(user, collection_id)
-        if collection is None:
-            return None
-
-        # Delete collection and release quota in a single transaction
-        async def _delete_collection_with_quota(session):
-            from sqlalchemy import select
-
-            from aperag.db.models import CollectionStatus
-            from aperag.service.quota_service import quota_service
-            from aperag.utils.utils import utc_now
-
-            # Get collection within transaction
-            stmt = select(db_models.Collection).where(
-                db_models.Collection.id == collection_id, db_models.Collection.user == user
-            )
-            result = await session.execute(stmt)
-            collection_to_delete = result.scalars().first()
-
-            if not collection_to_delete:
-                return None
-
-            # Mark collection as deleted
-            collection_to_delete.status = CollectionStatus.DELETED
-            collection_to_delete.gmt_deleted = utc_now()
-
-            # Release quota within the same transaction
-            await quota_service.release_quota(user, "max_collection_count", 1, session)
-
-            await session.flush()
-            await session.refresh(collection_to_delete)
-
-            return collection_to_delete
-
-        deleted_instance = await self.db_ops.execute_with_transaction(_delete_collection_with_quota)
-
-        if deleted_instance:
-            # Clean up related resources
-            collection_delete_task.delay(collection_id)
-            return await self.build_collection_response(deleted_instance)
-
-        return None
-
-    async def execute_search_flow(
-        self,
-        data: view_models.SearchRequest,
-        collection_id: str,
-        search_user_id: str,
-        chat_id: Optional[str] = None,
-        flow_name: str = "search",
-        flow_title: str = "Search",
-    ) -> Tuple[List[SearchResultItem], str]:
-        """Execute search using direct Python orchestration."""
-        _ = (flow_name, flow_title)
-        return await search_pipeline_service.execute_search(
-            data=data,
-            collection_id=collection_id,
-            search_user_id=search_user_id,
-            chat_id=chat_id,
-        )
-
-    async def create_search(
-        self, user: str, collection_id: str, data: view_models.SearchRequest
-    ) -> view_models.SearchResult:
-        from aperag.exceptions import CollectionNotFoundException
-
-        # Try to find collection as owner first
-        collection = await self.db_ops.query_collection(user, collection_id)
-        search_user_id = user  # Default to current user for search operations
-
-        if not collection:
-            # If not found as owner, check if it's a marketplace collection
-            try:
-                marketplace_info = await marketplace_collection_service._check_marketplace_access(user, collection_id)
-                # Use owner's user_id for search operations in marketplace collections
-                search_user_id = marketplace_info["owner_user_id"]
-                collection = await self.db_ops.query_collection(search_user_id, collection_id)
-                if not collection:
-                    raise CollectionNotFoundException(collection_id)
-            except Exception:
-                # If marketplace access also fails, raise original not found error
-                raise CollectionNotFoundException(collection_id)
-
-        # Execute search flow using helper method
-        items, _ = await self.execute_search_flow(
-            data=data,
-            collection_id=collection_id,
-            search_user_id=search_user_id,
-            chat_id=None,  # No chat filtering for regular collection searches
-            flow_name="search",
-            flow_title="Search",
-        )
-
-        # Save to database only if save_to_history is True
-        if data.save_to_history:
-            record = await self.db_ops.create_search(
-                user=user,
-                collection_id=collection_id,
-                query=data.query,
-                vector_search=data.vector_search.model_dump() if data.vector_search else None,
-                fulltext_search=data.fulltext_search.model_dump() if data.fulltext_search else None,
-                graph_search=data.graph_search.model_dump() if data.graph_search else None,
-                summary_search=data.summary_search.model_dump() if data.summary_search else None,
-                vision_search=data.vision_search.model_dump() if data.vision_search else None,
-                items=[item.model_dump() for item in items],
-            )
-            return SearchResult(
-                id=record.id,
-                query=record.query,
-                vector_search=record.vector_search,
-                fulltext_search=record.fulltext_search,
-                graph_search=record.graph_search,
-                summary_search=record.summary_search,
-                vision_search=record.vision_search,
-                items=items,
-                created=record.gmt_created.isoformat(),
-            )
-        else:
-            # Return search result without saving to database
-            return SearchResult(
-                id=None,  # No ID since not saved
-                query=data.query,
-                vector_search=data.vector_search,
-                fulltext_search=data.fulltext_search,
-                graph_search=data.graph_search,
-                summary_search=data.summary_search,
-                vision_search=data.vision_search,
-                items=items,
-                created=None,  # No creation time since not saved
-            )
-
-    async def list_searches(self, user: str, collection_id: str) -> view_models.SearchResultList:
-        from aperag.exceptions import CollectionNotFoundException
-
-        collection = await self.db_ops.query_collection(user, collection_id)
-        if not collection:
-            raise CollectionNotFoundException(collection_id)
-
-        # Use DatabaseOps to query searches
-        searches = await self.db_ops.query_searches(user, collection_id)
-
-        items = []
-        for search in searches:
-            search_result_items = []
-            for item_data in search.items:
-                search_result_items.append(SearchResultItem(**item_data))
-            items.append(
-                SearchResult(
-                    id=search.id,
-                    query=search.query,
-                    vector_search=search.vector_search,
-                    fulltext_search=search.fulltext_search,
-                    graph_search=search.graph_search,
-                    summary_search=search.summary_search,
-                    items=search_result_items,
-                    created=search.gmt_created.isoformat(),
-                )
-            )
-        return SearchResultList(items=items)
-
-    async def delete_search(self, user: str, collection_id: str, search_id: str) -> Optional[bool]:
-        """Delete search by ID (idempotent operation)
-
-        Returns True if deleted, None if already deleted/not found
-        """
-        from aperag.exceptions import CollectionNotFoundException
-
-        collection = await self.db_ops.query_collection(user, collection_id)
-        if not collection:
-            raise CollectionNotFoundException(collection_id)
-
-        return await self.db_ops.delete_search(user, collection_id, search_id)
-
-    async def validate_collections_batch(
-        self, user: str, collections: list[view_models.Collection]
-    ) -> tuple[bool, str]:
-        """
-        Validate multiple collections in a single database call.
-
-        Args:
-            user: User identifier
-            collections: List of collection objects to validate
-
-        Returns:
-            Tuple of (is_valid, error_message). If valid, error_message is empty.
-        """
-        if not collections:
-            return True, ""
-
-        # Extract collection IDs and validate they exist
-        collection_ids = []
-        for collection in collections:
-            if not collection.id:
-                return False, "Collection object missing 'id' field"
-            collection_ids.append(collection.id)
-
-        # Remove duplicates while preserving order
-        unique_collection_ids = list(dict.fromkeys(collection_ids))
-
-        try:
-            # Single database call to get all collections
-            db_collections = await self.db_ops.query_collections_by_ids(user, unique_collection_ids)
-
-            # Create a set of found collection IDs for fast lookup
-            found_collection_ids = {str(col.id) for col in db_collections}
-
-            # Check if all requested collections were found
-            for collection_id in unique_collection_ids:
-                if collection_id not in found_collection_ids:
-                    return False, f"Collection {collection_id} not found"
-
-            return True, ""
-
-        except Exception as e:
-            return False, f"Failed to validate collections: {str(e)}"
-
-    async def test_mineru_token(self, token: str) -> dict:
-        """Test the MinerU API token."""
-        async with httpx.AsyncClient() as client:
-            try:
-                response = await client.get(
-                    "https://mineru.net/api/v4/extract-results/batch/test-token",
-                    headers={"Authorization": f"Bearer {token}"},
-                )
-                return {"status_code": response.status_code, "data": response.json()}
-            except httpx.RequestError as e:
-                return {"status_code": 500, "data": {"msg": f"Request failed: {e}"}}
-
-
-# Create a global service instance for easy access
-# This uses the global db_ops instance and doesn't require session management in views
-collection_service = CollectionService()
+_bootstrap_knowledge_base_collection_service_di()
